add table checks

+++ src/main/cpp/player.hpp
@@ -3,9 +3,9 @@
 #include "seasocks/WebSocket.h"
 
 class Player {
-	public:
-		static constexpr size_t SIZE = 10;
-		seasocks::WebSocket* connection;
-		int table[SIZE * SIZE];
-	private:
+    public:
+        static constexpr int SIZE = 10;
+        seasocks::WebSocket* connection;
+        int table[SIZE * SIZE];
+    private:
 };
